Новые хуки для уменьшения бойлерплейта
