+++ frontend/src/js/hooks/use-authorized-only.ts
@@ -1,0 +1,15 @@
+import { useSelector } from 'react-redux';
+import { isAuthorizedSelector } from '#src/js/redux/selectors';
+import { useEffect } from 'react';
+import { useHistory } from 'react-router-dom';
+
+export const useAuthorizedOnly = (): void => {
+  const isAuthorized = useSelector(isAuthorizedSelector);
+  const history = useHistory();
+
+  useEffect(() => {
+    if (!isAuthorized) {
+      history.replace(`/`);
+    }
+  }, [history, isAuthorized]);
+};
