+++ frontend/src/js/hooks/use-show-user-request-error.ts
@@ -1,0 +1,17 @@
+import { useEffect } from 'react';
+import UIkit from 'uikit';
+import { useSelector } from 'react-redux';
+import { userRequestErrorSelector } from '#src/js/redux/selectors';
+
+export const useShowUserRequestError = (isSubmitSuccessful: boolean): void => {
+  const userRequestError = useSelector(userRequestErrorSelector);
+
+  useEffect(() => {
+    if (isSubmitSuccessful && userRequestError)
+      UIkit.notification({
+        message: userRequestError.message,
+        pos: `bottom-right`,
+      });
+
+  }, [userRequestError, isSubmitSuccessful]);
+};
